rebuild_vector_db/requirements.txt
# Requirements for rebuild_vector_db module

# Text processing
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Vector database
chromadb>=0.4.0

# HTTP requests
requests>=2.31.0

# Progress bars
tqdm>=4.66.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
hypothesis>=6.92.0
